Introduce explicit classes DiscreteImage{1D,2D,3D} to avoid typeclass clutter
