+++ src/main/scala/scalismo/image/Image.scala
@@ -29,7 +29,7 @@ import DiscreteImageDomain.CanCreate
 /**
  * An image whose values are scalar.
  */
-class ScalarImage[D <: Dim: NDSpace: CanInterpolate] protected (override val domain: Domain[D], override val f: Point[D] => Float) extends ScalarField[D, Float](domain, f) {
+class ScalarImage[D <: Dim: NDSpace] protected (override val domain: Domain[D], override val f: Point[D] => Float) extends ScalarField[D, Float](domain, f) {
 
   /** adds two images. The domain of the new image is the intersection of both */
   def +(that: ScalarImage[D]): ScalarImage[D] = {
@@ -108,7 +108,7 @@ class ScalarImage[D <: Dim: NDSpace: CanInterpolate] protected (override val dom
    * Returns a discrete scalar image with the given domain, whose values are obtained by sampling the scalarImge at the domain points.
    * If the image is not defined at a domain point, the outside value is used.
    */
-  def sample[Pixel: Scalar: ClassTag](domain: DiscreteImageDomain[D], outsideValue: Float): DiscreteScalarImage[D, Pixel] = {
+  def sample[Pixel: Scalar: ClassTag](domain: DiscreteImageDomain[D], outsideValue: Float)(implicit ev: DiscreteScalarImage.Create[D]): DiscreteScalarImage[D, Pixel] = {
     val numeric = implicitly[Scalar[Pixel]]
 
     val sampledValues = domain.points.toIterable.par.map((pt: Point[D]) => {
@@ -132,14 +132,14 @@ object ScalarImage {
    * @param domain The domain over which the image is defined
    * @param f A function which yields for each point of the domain its value
    */
-  def apply[D <: Dim: NDSpace: CanInterpolate](domain: Domain[D], f: Point[D] => Float) = new ScalarImage[D](domain, f)
+  def apply[D <: Dim: NDSpace](domain: Domain[D], f: Point[D] => Float) = new ScalarImage[D](domain, f)
 
 }
 
 /**
  * A scalar image that is once differentiable
  */
-class DifferentiableScalarImage[D <: Dim: NDSpace: CanInterpolate](_domain: Domain[D], _f: Point[D] => Float, val df: Point[D] => Vector[D]) extends ScalarImage[D](_domain, _f) {
+class DifferentiableScalarImage[D <: Dim: NDSpace](_domain: Domain[D], _f: Point[D] => Float, val df: Point[D] => Vector[D]) extends ScalarImage[D](_domain, _f) {
 
   def differentiate: VectorField[D, D] = VectorField(domain, df)
 
@@ -216,7 +216,7 @@ object DifferentiableScalarImage {
    * @param f a function that yiels for each point of the domain its intensities
    * @param df the derivative of the function f
    */
-  def apply[D <: Dim: NDSpace: CanInterpolate](domain: Domain[D], f: Point[D] => Float, df: Point[D] => Vector[D]) = new DifferentiableScalarImage[D](domain, f, df)
+  def apply[D <: Dim: NDSpace](domain: Domain[D], f: Point[D] => Float, df: Point[D] => Vector[D]) = new DifferentiableScalarImage[D](domain, f, df)
 
 }
