+++ src/main/scala/scalismo/image/filter/DiscreteImageFilter.scala
@@ -17,7 +17,7 @@ package scalismo.image.filter
 
 import scalismo.common.{ ScalarArray, Scalar }
 import scalismo.common.DiscreteDomain.CanBound
-import scalismo.image.{ CanInterpolate, DiscreteScalarImage }
+import scalismo.image.{ DiscreteScalarImage }
 import scalismo.geometry._
 import scalismo.utils.{ CanConvertToVtk, ImageConversion }
 import vtk.{ vtkImageCast, vtkImageEuclideanDistance }
@@ -31,7 +31,7 @@ object DiscreteImageFilter {
    * Computes a (signed) distance transform of the image.
    * @note The value that is returned is not the euclidean distance unless the image has unit spacing. Even worse, the distance might depend on the spacing of the image.
    */
-  def distanceTransform[D <: Dim: NDSpace: CanConvertToVtk: CanInterpolate, A: Scalar: ClassTag: TypeTag](img: DiscreteScalarImage[D, A]): DiscreteScalarImage[D, Float] = {
+  def distanceTransform[D <: Dim: NDSpace: CanConvertToVtk: DiscreteScalarImage.Create, A: Scalar: ClassTag: TypeTag](img: DiscreteScalarImage[D, A]): DiscreteScalarImage[D, Float] = {
 
     val scalar = implicitly[Scalar[A]]
